バックエンドとの接続確認
